Port the original `reproductions-3.cy.spec.js`
