--- e2e/test/scenarios/question/reproductions/33079-drill-underlying-records.cy.spec.js
@@ -1,34 +1,0 @@
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { popover, restore, cartesianChartCircle } from "e2e/support/helpers";
-const { ORDERS, ORDERS_ID } = SAMPLE_DATABASE;
-
-const questionDetails = {
-  display: "line",
-  query: {
-    "source-table": ORDERS_ID,
-    aggregation: [["count"]],
-    breakout: [["field", ORDERS.CREATED_AT, { "temporal-unit": "month" }]],
-  },
-};
-
-describe("issue 33079", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-
-    cy.intercept("POST", "/api/dataset").as("dataset");
-    cy.request("GET", "/api/user/current").then(({ body: user }) => {
-      cy.request("PUT", `/api/user/${user.id}`, { locale: "de" });
-    });
-  });
-
-  it("underlying records drill should work in a non-English locale (metabase#33079)", () => {
-    cy.createQuestion(questionDetails, { visitQuestion: true });
-    cartesianChartCircle().eq(1).click({ force: true });
-    popover()
-      .findByText(/Order/) // See these Orders
-      .click();
-    cy.wait("@dataset");
-    cy.findByTestId("question-row-count").should("contain", "19");
-  });
-});
