--- e2e/test/scenarios/question/reproductions/39795-reorder-column-field-mapping.cy.spec.js
@@ -1,42 +1,0 @@
-import { SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import {
-  getDraggableElements,
-  moveColumnDown,
-  visitQuestionAdhoc,
-  restore,
-} from "e2e/support/helpers";
-
-const { ORDERS, ORDERS_ID, PRODUCTS } = SAMPLE_DATABASE;
-
-describe("issue 39795", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-
-    //If you comment out this post, then the test will pass.
-    cy.request("post", `/api/field/${ORDERS.PRODUCT_ID}/dimension`, {
-      human_readable_field_id: PRODUCTS.TITLE,
-      name: "Product ID",
-      type: "external",
-    });
-  });
-
-  it("should allow me to re-order even when a field is set with a different display value (metabase#39795)", () => {
-    visitQuestionAdhoc({
-      dataset_query: {
-        database: SAMPLE_DB_ID,
-        query: {
-          "source-table": ORDERS_ID,
-        },
-        type: "query",
-      },
-    });
-    cy.findByTestId("viz-settings-button").click();
-    moveColumnDown(getDraggableElements().first(), 2);
-
-    // We are not able to re-order because the dataset will also contain values a column for Product ID
-    // This causes the isValid() check to fire, and you are always forced into the default value for table.columns
-    getDraggableElements().eq(2).should("contain.text", "ID");
-  });
-});
